JBPM-5667: Second stage of a Case verifies its completion condition with old value of case file property
